src/requirements.txt
celery==5.4.0
Django==4.2.13
django-celery-beat==2.6.0
django-jazzmin==3.0.0
djangorestframework==3.15.1
pybit==5.7.0
pycryptodome==3.20.0
pyTelegramBotAPI==4.18.1
redis==5.0.5
requests==2.31.0